Move ConfigurationContext to org.wso2.testgrid.common.config package
