+++ dao/src/main/java/org/wso2/testgrid/dao/EntityManagerHelper.java
@@ -20,8 +20,8 @@ package org.wso2.testgrid.dao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.testgrid.common.TestGridConstants;
-import org.wso2.testgrid.common.util.ConfigurationContext;
-import org.wso2.testgrid.common.util.ConfigurationContext.ConfigurationProperties;
+import org.wso2.testgrid.common.config.ConfigurationContext;
+import org.wso2.testgrid.common.config.ConfigurationContext.ConfigurationProperties;
 import org.wso2.testgrid.common.util.StringUtil;
 
 import java.util.HashMap;
